requirements-dev.txt
apache-airflow==2.3.3
apache-airflow-providers-cncf-kubernetes
pre-commit
pylint
pylint-airflow
